Put all classes in duke package
