+++ src/main/java/duke/AddCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class AddCommand extends Command {
     private Task task;
