+++ src/main/java/duke/DeleteCommand.java
@@ -1,20 +1,22 @@
-public class DoneCommand extends Command {
+package duke;
+
+public class DeleteCommand extends Command {
     private String index;
 
-    public DoneCommand(String index) {
+    public DeleteCommand(String index) {
         this.index = index;
     }
 
     /**
      * Executes the specific actions for this command.
      *
-     * @param tasks Handles the list of tasks.
-     * @param ui Handles the user interface.
+     * @param tasks   Handles the list of tasks.
+     * @param ui      Handles the user interface.
      * @param storage Handles the saving and loading of tasks.
      */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-        tasks.doneTask(storage, this.index);
+        tasks.deleteTask(storage, this.index);
     }
 
     /**
