+++ src/main/java/duke/Duke.java
@@ -1,14 +1,4 @@
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Scanner;
-import java.util.ArrayList;
-import java.util.List;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.FileWriter;
+package duke;
 
 /**
  * Represents a personal assistant chatbot that responds to command line inputs.
