+++ src/main/java/duke/Todo.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Represents a task without any date/time attached to it.
  *
